TopOptProblems

This sub-module of TopOpt defines a number of standard topology optimization problems for the convenient testing of algorithms.

Problem types

Abstract type

StiffnessTopOptProblem is an abstract type that a number of linear elasticity, quasi-static, topology optimization problems subtype.

TopOpt.TopOptProblems.StiffnessTopOptProblem — Type

Abstract stiffness topology optimization problem. All subtypes must have the following fields: ch::ConstraintHandler black::BitVector white::BitVector varind::AbstractVector{Int} metadata::Metadata

Concrete types

The following types are all concrete subtypes of StiffnessTopOptProblem. PointLoadCantilever is a cantilever beam problem with a point load as shown below. HalfMBB is the half Messerschmitt-Bölkow-Blohm (MBB) beam problem commonly used in topology optimization literature. LBeam and TieBeam are the common L-beam and tie-beam test problem used in topology optimization literature. The PointLoadCantilever and HalfMBB problems can be either 2D or 3D depending on the type of the inputs to the constructor. If the number of elements and sizes of elements are 2-tuples, the problem constructed will be 2D. And if they are 3-tuples, the problem constructed will be 3D. For the 3D versions, the point loads are applied at approximately the mid-depth point. The TieBeam and LBeam problems are always 2D.

Finally, InpStiffness is a problem type that is instantiated by importing an .inp file. This can be used to represent an arbitrary unstructured mesh with complex boundary condition domains and load specification. The .inp file can be exported from a number of common finite element software such as: FreeCAD or ABAQUS.

TopOpt.TopOptProblems.PointLoadCantilever — Type

///**********************************
///*                                *
///*                                * |
///*                                * |
///********************************** v

struct PointLoadCantilever{dim, T, N, M} <: StiffnessTopOptProblem{dim, T}
    rect_grid::RectilinearGrid{dim, T, N, M}
    E::T
    ν::T
    ch::ConstraintHandler{DofHandler{dim, N, T, M}, T}
    force::T
    force_dof::Int
    metadata::Metadata
end

API:

    PointLoadCantilever(::Type{Val{CellType}}, nels::NTuple{dim,Int}, sizes::NTuple{dim}, E, ν, force) where {dim, CellType}

dim: dimension of the problem

T: number type for computations and coordinates

N: number of nodes in a cell of the grid

M: number of faces in a cell of the grid

rect_grid: a RectilinearGrid struct

E: Young's modulus

ν: Poisson's ration

ch: a JuAFEM.ConstraintHandler struct

force: force at the center right of the cantilever beam (positive is downward)

force_dof: dof number at which the force is applied

metadata: Metadata having various cell-node-dof relationships

nels: number of elements in each direction, a 2-tuple for 2D problems and a 3-tuple for 3D problems

sizes: the size of each element in each direction, a 2-tuple for 2D problems and a 3-tuple for 3D problems

CellType: can be either :Linear or :Quadratic to determine the order of the geometric and field basis functions and element type. Only isoparametric elements are supported for now.

Example:

nels = (60,20);
sizes = (1.0,1.0);
E = 1.0;
ν = 0.3;
force = 1.0;

# Linear elements and linear basis functions
celltype = :Linear

# Quadratic elements and quadratic basis functions
#celltype = :Quadratic

problem = PointLoadCantilever(Val{celltype}, nels, sizes, E, ν, force)

TopOpt.TopOptProblems.HalfMBB — Type

 |
 |
 v
O*********************************
O*                               *
O*                               *
O*                               *
O*********************************
                                 O

struct HalfMBB{dim, T, N, M} <: StiffnessTopOptProblem{dim, T}
    rect_grid::RectilinearGrid{dim, T, N, M}
    E::T
    ν::T
    ch::ConstraintHandler{DofHandler{dim, N, T, M}, T}
    force::T
    force_dof::Int
    metadata::Metadata
end

API:

    HalfMBB(::Type{Val{CellType}}, nels::NTuple{dim,Int}, sizes::NTuple{dim}, E, ν, force) where {dim, CellType}

dim: dimension of the problem

T: number type for computations and coordinates

N: number of nodes in a cell of the grid

M: number of faces in a cell of the grid

rect_grid: a RectilinearGrid struct

E: Young's modulus

ν: Poisson's ration

ch: a JuAFEM.ConstraintHandler struct

force: force at the top left of half the MBB (positive is downward)

force_dof: dof number at which the force is applied

metadata: Metadata having various cell-node-dof relationships

nels: number of elements in each direction, a 2-tuple for 2D problems and a 3-tuple for 3D problems

sizes: the size of each element in each direction, a 2-tuple for 2D problems and a 3-tuple for 3D problems

CellType: can be either :Linear or :Quadratic to determine the order of the geometric and field basis functions and element type. Only isoparametric elements are supported for now.

Example:

nels = (60,20);
sizes = (1.0,1.0);
E = 1.0;
ν = 0.3;
force = -1.0;

# Linear elements and linear basis functions
celltype = :Linear

# Quadratic elements and quadratic basis functions
#celltype = :Quadratic

problem = HalfMBB(Val{celltype}, nels, sizes, E, ν, force)

TopOpt.TopOptProblems.LBeam — Type

////////////
............
.          .
.          .
.          . 
.          .                    
.          ......................
.                               .
.                               . 
.                               . |
................................. v
                                force

struct LBeam{T, N, M} <: StiffnessTopOptProblem{2, T}
    E::T
    ν::T
    ch::ConstraintHandler{<:DofHandler{2, N, T, M}, T}
    force::T
    force_dof::Integer
    black::AbstractVector
    white::AbstractVector
    varind::AbstractVector{Int}
    metadata::Metadata
end

API:

    LBeam(::Type{Val{CellType}}, ::Type{T}=Float64; length = 100, height = 100, upperslab = 50, lowerslab = 50, E = 1.0, ν = 0.3, force = 1.0) where {T, CellType}

T: number type for computations and coordinates

N: number of nodes in a cell of the grid

M: number of faces in a cell of the grid

E: Young's modulus

ν: Poisson's ration

ch: a JuAFEM.ConstraintHandler struct

force: force at the center right of the cantilever beam (positive is downward)

force_dof: dof number at which the force is applied

metadata:: Metadata having various cell-node-dof relationships

length, height, upperslab and lowerslab are explained in LGrid.

CellType: can be either :Linear or :Quadratic to determine the order of the geometric and field basis functions and element type. Only isoparametric elements are supported for now.

Example:

E = 1.0;
ν = 0.3;
force = 1.0;

# Linear elements and linear basis functions
celltype = :Linear

# Quadratic elements and quadratic basis functions
#celltype = :Quadratic

problem = LBeam(Val{celltype}, E = E, ν = ν, force = force)

TopOpt.TopOptProblems.TieBeam — Type

                                                               1
                                                               
                                                              OOO
                                                              ...
                                                              . .
                                                           4  . . 
                                30                            . .   
/ . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . <-
/ .                                                                 . <- 2 f 
/ .    3                                                            . <- 
/ . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . <-
                                                              ^^^
                                                              |||
                                                              1 f

TopOpt.TopOptProblems.IO.INP.InpStiffness — Type

InpStiffness(filepath::AbstractString; keep_load_cells = false)

Imports stiffness problem from a .inp file, e.g. InpStiffness("example.inp"). The keep_load_cells keyword argument will enforce that any cell with a load applied on it is to be part of the final optimal design generated by topology optimization algorithms.

Grids

Grid types are defined in TopOptProblems because a number of topology optimization problems share the same underlying grid but apply the loads and boundary conditions at different locations. For example, the PointLoadCantilever and HalfMBB problems use the same rectilinear grid, RectilinearGrid, under the hood. The LBeam problem uses the LGrid grid type under the hood. New problem types can be defined using the same grid types but different loads or boundary conditions.

TopOpt.TopOptProblems.RectilinearGrid — Type

struct RectilinearGrid{dim, T, N, M} <: AbstractGrid{dim, T}
    grid::JuAFEM.Grid{dim, N, T, M}
    nels::NTuple{dim, Int}
    sizes::NTuple{dim, T}
    corners::NTuple{2, Vec{dim,T}}
    white_cells::BitVector
    black_cells::BitVector
    constant_cells::BitVector
end

dim: dimension of the problem

T: number type for computations and coordinates

N: number of nodes in a cell of the grid

M: number of faces in a cell of the grid

grid: a JuAFEM.Grid struct

nels: number of elements in every dimension

sizes: dimensions of each rectilinear cell

corners: 2 corner points of the rectilinear grid

white_cells: cells fixed to be void during optimization

black_cells: cells fixed to have material during optimization

constant_cells: cells fixed to be either void or have material during optimization

API:

function RectilinearGrid(nels::NTuple{dim,Int}, sizes::NTuple{dim,T}) where {dim,T}

Example:

rectgrid = RectilinearGrid((60,20), (1.0,1.0))

TopOpt.TopOptProblems.LGrid — Function

Examples:

grid = LGrid(upperslab = 30, lowerslab = 70)
grid = LGrid(Val{:Linear}, (2, 4), (2, 2), Vec{2,Float64}((0.0,0.0)), Vec{2,Float64}((2.0, 4.0)), Vec{2,Float64}((4.0, 2.0)))

        upperslab
       ............
       .          .
       .          .
       .          . 
height .          .                    
       .          ......................
       .                               .
       .                               . lowerslab
       .                               .
       .................................
                    length

            UR
 ............
 .          .
 .          .
 .          . 
 .          .                    MR
 .          ......................
 .                               .
 .                               .
 .                               .
 .................................
LL

Finite element backend

Currently, TopOpt uses a forked version of JuAFEM.jl. This means that all the problems above are described in the language and types of JuAFEM.

The fork is needed for GPU support but the main package should also work on the CPU. The changes in the fork should make it back to the main repo at some point.

Element matrices and vectors

ElementFEAInfo

TopOpt.TopOptProblems.ElementFEAInfo — Type

struct ElementFEAInfo{dim, T}
    Kes::AbstractVector{<:AbstractMatrix{T}}
    fes::AbstractVector{<:AbstractVector{T}}
    fixedload::AbstractVector{T}
    cellvolumes::AbstractVector{T}
    cellvalues::CellValues{dim, T}
    facevalues::FaceValues{<:Any, T}
    metadata::Metadata
    black::AbstractVector
    white::AbstractVector
    varind::AbstractVector{Int}
    cells
end

An instance of the ElementFEAInfo type stores element information such as:

Kes: the element stiffness matrices,

fes: the element load vectors,

cellvolumes: the element volumes,

cellvalues and facevalues: two JuAFEM types that facilitate cell and face iteration and queries.

metadata: that stores degree of freedom (dof) to node mapping, dof to cell mapping, etc.

black: a BitVector such that black[i] is 1 iff element i must be part of any feasible design.

white: a BitVector such that white[i] is 1 iff element i must never be part of any feasible design.

varind: a vector such that varind[i] gives the decision variable index of element i.

cells: the cell connectivities,

GlobalFEAInfo

TopOpt.TopOptProblems.GlobalFEAInfo — Type

struct GlobalFEAInfo{T, TK<:AbstractMatrix{T}, Tf<:AbstractVector{T}, Tchol}
    K::TK
    f::Tf
    cholK::Tchol
end

An instance of GlobalFEAInfo hosts the global stiffness matrix K, the load vector f and the cholesky decomposition of the K, cholK.