TopOptProblems

This sub-module of TopOpt defines a number of standard topology optimization problems for the convenient testing of algorithms.

Problem types

Abstract type

StiffnessTopOptProblem is an abstract type that a number of linear elasticity, quasi-static, topology optimization problems subtype.

TopOpt.TopOptProblems.StiffnessTopOptProblem — Type

Abstract stiffness topology optimization problem. All subtypes must have the following fields: ch::ConstraintHandler black::BitVector white::BitVector varind::AbstractVector{Int} metadata::Metadata

Concrete types

The following types are all concrete subtypes of StiffnessTopOptProblem. PointLoadCantilever is a cantilever beam problem with a point load as shown below. HalfMBB is the half Messerschmitt-Bölkow-Blohm (MBB) beam problem commonly used in topology optimization literature. LBeam and TieBeam are the common L-beam and tie-beam test problem used in topology optimization literature. The PointLoadCantilever and HalfMBB problems can be either 2D or 3D depending on the type of the inputs to the constructor. If the number of elements and sizes of elements are 2-tuples, the problem constructed will be 2D. And if they are 3-tuples, the problem constructed will be 3D. For the 3D versions, the point loads are applied at approximately the mid-depth point. The TieBeam and LBeam problems are always 2D.

Finally, InpStiffness is a problem type that is instantiated by importing an .inp file. This can be used to represent an arbitrary unstructured mesh with complex boundary condition domains and load specification. The .inp file can be exported from a number of common finite element software such as: FreeCAD or ABAQUS.

TopOpt.TopOptProblems.PointLoadCantilever — Type

///**********************************
///*                                *
///*                                * |
///*                                * |
///********************************** v

struct PointLoadCantilever{dim, T, N, M} <: StiffnessTopOptProblem{dim, T}
    rect_grid::RectilinearGrid{dim, T, N, M}
    E::T
    ν::T
    ch::ConstraintHandler{DofHandler{dim, N, T, M}, T}
    force::T
    force_dof::Int
    metadata::Metadata
end

dim: dimension of the problem

T: number type for computations and coordinates

N: number of nodes in a cell of the grid

M: number of faces in a cell of the grid

rect_grid: a RectilinearGrid struct

E: Young's modulus

ν: Poisson's ration

ch: a JuAFEM.ConstraintHandler struct

force: force at the center right of the cantilever beam (positive is downward)

force_dof: dof number at which the force is applied

metadata:: Metadata having various cell-node-dof relationships

API:

    PointLoadCantilever(nels::NTuple{dim,Int}, sizes::NTuple{dim}, E, ν, force) where {dim}

Example:

nels = (60,20);
sizes = (1.0,1.0);
E = 1.0;
ν = 0.3;
force = 1.0;
problem = PointLoadCantilever(nels, sizes, E, ν, force)

TopOpt.TopOptProblems.HalfMBB — Type

 |
 |
 v
O*********************************
O*                               *
O*                               *
O*                               *
O*********************************
                                 O

struct HalfMBB{dim, T, N, M} <: StiffnessTopOptProblem{dim, T}
    rect_grid::RectilinearGrid{dim, T, N, M}
    E::T
    ν::T
    ch::ConstraintHandler{DofHandler{dim, N, T, M}, T}
    force::T
    force_dof::Int
    metadata::Metadata
end

dim: dimension of the problem

T: number type for computations and coordinates

N: number of nodes in a cell of the grid

M: number of faces in a cell of the grid

rect_grid: a RectilinearGrid struct

E: Young's modulus

ν: Poisson's ration

ch: a JuAFEM.ConstraintHandler struct

force: force at the top left of half the MBB (positive is downward)

force_dof: dof number at which the force is applied

metadata:: Metadata having various cell-node-dof relationships

API:

    HalfMBB(nels::NTuple{dim,Int}, sizes::NTuple{dim}, E, ν, force) where {dim}

Example:

nels = (60,20);
sizes = (1.0,1.0);
E = 1.0;
ν = 0.3;
force = -1.0;
problem = HalfMBB(nels, sizes, E, ν, force)

TopOpt.TopOptProblems.LBeam — Type

////////////
............
.          .
.          .
.          . 
.          .                    
.          ......................
.                               .
.                               . 
.                               . |
................................. v
                                force

TopOpt.TopOptProblems.TieBeam — Type

                                                               1
                                                               
                                                              OOO
                                                              ...
                                                              . .
                                                           4  . . 
                                30                            . .   
/ . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . <-
/ .                                                                 . <- 2 f 
/ .    3                                                            . <- 
/ . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . <-
                                                              ^^^
                                                              |||
                                                              1 f

TopOpt.TopOptProblems.IO.INP.InpStiffness — Type

Stiffness problem imported from a .inp file.

Grids

Grid types are defined in TopOptProblems because a number of topology optimization problems share the same underlying grid but apply the loads and boundary conditions at different locations. For example, the PointLoadCantilever and HalfMBB problems use the same rectilinear grid, RectilinearGrid, under the hood. The LBeam problem uses the LGrid grid type under the hood. New problem types can be defined using the same grid types but different loads or boundary conditions.

TopOpt.TopOptProblems.RectilinearGrid — Type

struct RectilinearGrid{dim, T, N, M} <: AbstractGrid{dim, T}
    grid::JuAFEM.Grid{dim, N, T, M}
    nels::NTuple{dim, Int}
    sizes::NTuple{dim, T}
    corners::NTuple{2, Vec{dim,T}}
    white_cells::BitVector
    black_cells::BitVector
    constant_cells::BitVector
end

dim: dimension of the problem

T: number type for computations and coordinates

N: number of nodes in a cell of the grid

M: number of faces in a cell of the grid

grid: a JuAFEM.Grid struct

nels: number of elements in every dimension

sizes: dimensions of each rectilinear cell

corners: 2 corner points of the rectilinear grid

white_cells: cells fixed to be void during optimization

black_cells: cells fixed to have material during optimization

constant_cells: cells fixed to be either void or have material during optimization

API:

function RectilinearGrid(nels::NTuple{dim,Int}, sizes::NTuple{dim,T}) where {dim,T}

Example:

rectgrid = RectilinearGrid((60,20), (1.0,1.0))

TopOpt.TopOptProblems.LGrid — Function

Examples:

grid = LGrid(upperslab = 30, lowerslab = 70)
grid = LGrid(Val{:Linear}, (2, 4), (2, 2), Vec{2,Float64}((0.0,0.0)), Vec{2,Float64}((2.0, 4.0)), Vec{2,Float64}((4.0, 2.0)))

        upperslab
       ............
       .          .
       .          .
       .          . 
height .          .                    
       .          ......................
       .                               .
       .                               . lowerslab
       .                               .
       .................................
                    length

            UR
 ............
 .          .
 .          .
 .          . 
 .          .                    MR
 .          ......................
 .                               .
 .                               .
 .                               .
 .................................
LL

Finite element backend

Currently, TopOpt uses a forked version of JuAFEM.jl. This means that all the problems above are described in the language and types of JuAFEM.

The fork is needed for GPU support but the main package should also work on the CPU. The changes in the fork should make it back to the main repo at some point.

Element matrices and vectors

Generating the element stiffness matrices and load vectors from a linear elasticity stiffness problem sp::StiffnessTopOptProblem using a Gaussian quadrature order quad_order is done using:

elementinfo = ElementFEAInfo(sp, quad_order)